+++ Audibly.App/Views/LibraryCardPage.xaml.cs
@@ -21,6 +21,7 @@ using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Input;
 using Sentry;
+using Sharpener.Extensions;
 
 namespace Audibly.App.Views;
 
@@ -29,7 +30,7 @@ namespace Audibly.App.Views;
 /// </summary>
 public sealed partial class LibraryCardPage : Page
 {
-    private readonly DispatcherQueue dispatcherQueue = DispatcherQueue.GetForCurrentThread();
+    private readonly DispatcherQueue _dispatcherQueue = DispatcherQueue.GetForCurrentThread();
 
     /// <summary>
     ///     Gets the app-wide ViewModel instance.
@@ -87,7 +88,7 @@ public sealed partial class LibraryCardPage : Page
 
                 // delete the old cover images
 
-                await dispatcherQueue.EnqueueAsync(() => ViewModel.IsLoading = true);
+                await _dispatcherQueue.EnqueueAsync(() => ViewModel.IsLoading = true);
 
                 ViewModel.MessageService.ShowDialog(DialogType.Progress, "Data Migration",
                     "Deleting old cover images");
@@ -96,7 +97,7 @@ public sealed partial class LibraryCardPage : Page
                     importedAudiobooks.Select(x => x.CoverImagePath).ToList(),
                     async (i, count, _) =>
                     {
-                        await dispatcherQueue.EnqueueAsync(() =>
+                        await _dispatcherQueue.EnqueueAsync(() =>
                         {
                             ViewModel.ProgressDialogProgress = ((double)i / count * 100).ToInt();
                             ViewModel.ProgressDialogPrefix = "Deleting audiobook";
@@ -104,7 +105,7 @@ public sealed partial class LibraryCardPage : Page
                         });
                     });
 
-                await dispatcherQueue.EnqueueAsync(() =>
+                await _dispatcherQueue.EnqueueAsync(() =>
                 {
                     // ViewModel.ProgressDialogText = string.Empty;
                     ViewModel.ProgressDialogProgress = 0;
@@ -121,7 +122,7 @@ public sealed partial class LibraryCardPage : Page
                 await ViewModel.FileImporter.ImportFromJsonAsync(file, new System.Threading.CancellationToken(),
                     async (i, count, title, _) =>
                     {
-                        await dispatcherQueue.EnqueueAsync(() =>
+                        await _dispatcherQueue.EnqueueAsync(() =>
                         {
                             ViewModel.ProgressDialogProgress = ((double)i / count * 100).ToInt();
                             ViewModel.ProgressDialogPrefix = "Importing";
@@ -131,7 +132,7 @@ public sealed partial class LibraryCardPage : Page
 
                 ViewModel.OnProgressDialogCompleted();
 
-                await dispatcherQueue.EnqueueAsync(() =>
+                await _dispatcherQueue.EnqueueAsync(() =>
                 {
                     ViewModel.ProgressDialogPrefix = string.Empty;
                     ViewModel.ProgressDialogText = string.Empty;
@@ -150,6 +151,16 @@ public sealed partial class LibraryCardPage : Page
 
                 await ViewModel.GetAudiobookListAsync(true);
             });
+    }
+
+    private async void RefreshButton_OnClick(object sender, RoutedEventArgs e)
+    {
+        // unchecked all the filter flyout items
+        InProgressFilterCheckBox.IsChecked = false;
+        NotStartedFilterCheckBox.IsChecked = false;
+        CompletedFilterCheckBox.IsChecked = false;
+
+        await ViewModel.GetAudiobookListAsync();
     }
 
     private void TestContentDialogButton_OnClick(object sender, RoutedEventArgs e)
@@ -237,5 +248,163 @@ public sealed partial class LibraryCardPage : Page
             Message = "Sentry message sent",
             Severity = InfoBarSeverity.Success
         });
+    }
+
+    public enum AudioBookFilter
+    {
+        InProgress,
+        NotStarted,
+        Completed
+    }
+
+    /// <summary>
+    ///     Resets the audiobook list.
+    /// </summary>
+    public async Task ResetAudiobookListAsync()
+    {
+        _activeFilters.Clear();
+
+        // unchecked all the filter flyout items
+        InProgressFilterCheckBox.IsChecked = false;
+        NotStartedFilterCheckBox.IsChecked = false;
+        CompletedFilterCheckBox.IsChecked = false;
+
+        await _dispatcherQueue.EnqueueAsync(() =>
+        {
+            ViewModel.Audiobooks.Clear();
+            foreach (var a in ViewModel.AudiobooksForFilter) ViewModel.Audiobooks.Add(a);
+        });
+    }
+
+    private HashSet<AudioBookFilter> _activeFilters = new();
+
+    private List<AudiobookViewModel> GetFilteredAudiobooks()
+    {
+        // matches audiobooks for each active filter
+        var matches = new List<AudiobookViewModel>();
+
+        foreach (var audiobook in ViewModel.AudiobooksForFilter)
+        {
+            if (_activeFilters.Contains(AudioBookFilter.InProgress) && audiobook.Progress > 0 && !audiobook.IsCompleted)
+                matches.Add(audiobook);
+            if (_activeFilters.Contains(AudioBookFilter.NotStarted) && audiobook.Progress == 0)
+                matches.Add(audiobook);
+            if (_activeFilters.Contains(AudioBookFilter.Completed) && audiobook.IsCompleted)
+                matches.Add(audiobook);
+        }
+
+        return matches;
+    }
+
+    /// <summary>
+    ///     Filters the audiobook list based on the search text.
+    /// </summary>
+    private async Task FilterAudiobookList()
+    {
+        if (_activeFilters.Count == 0)
+        {
+            await ResetAudiobookListAsync();
+            return;
+        }
+
+        var matches = GetFilteredAudiobooks();
+
+        await _dispatcherQueue.EnqueueAsync(() =>
+        {
+            ViewModel.Audiobooks.Clear();
+            foreach (var match in matches) ViewModel.Audiobooks.Add(match);
+        });
+    }
+
+    private void SetCheckedState()
+    {
+        // Controls are null the first time this is called, so we just 
+        // need to perform a null check on any one of the controls.
+        if (InProgressFilterCheckBox == null) return;
+        
+        if (InProgressFilterCheckBox.IsChecked == true &&
+            NotStartedFilterCheckBox.IsChecked == true &&
+            CompletedFilterCheckBox.IsChecked == true)
+            SelectAllFiltersCheckBox.IsChecked = true;
+        else if (InProgressFilterCheckBox.IsChecked == false &&
+                 NotStartedFilterCheckBox.IsChecked == false &&
+                 CompletedFilterCheckBox.IsChecked == false)
+            SelectAllFiltersCheckBox.IsChecked = false;
+        else
+            // Set third state (indeterminate) by setting IsChecked to null.
+            SelectAllFiltersCheckBox.IsChecked = null;
+    }
+
+    private async void InProgressFilterCheckBox_OnChecked(object sender, RoutedEventArgs e)
+    {
+        SetCheckedState();
+
+        _activeFilters.Add(AudioBookFilter.InProgress);
+
+        await FilterAudiobookList();
+    }
+
+    private async void NotStartedFilterCheckBox_OnChecked(object sender, RoutedEventArgs e)
+    {
+        SetCheckedState();
+
+        _activeFilters.Add(AudioBookFilter.NotStarted);
+
+        await FilterAudiobookList();
+    }
+
+    private async void CompletedFilterCheckBox_OnChecked(object sender, RoutedEventArgs e)
+    {
+        SetCheckedState();
+
+        _activeFilters.Add(AudioBookFilter.Completed);
+
+        await FilterAudiobookList();
+    }
+
+    private async void InProgressFilterCheckBox_OnUnchecked(object sender, RoutedEventArgs e)
+    {
+        SetCheckedState();
+
+        _activeFilters.Remove(AudioBookFilter.InProgress);
+
+        await FilterAudiobookList();
+    }
+
+    private async void NotStartedFilterCheckBox_OnUnchecked(object sender, RoutedEventArgs e)
+    {
+        SetCheckedState();
+
+        _activeFilters.Remove(AudioBookFilter.NotStarted);
+
+        await FilterAudiobookList();
+    }
+
+    private async void CompletedFilterCheckBox_OnUnchecked(object sender, RoutedEventArgs e)
+    {
+        SetCheckedState();
+
+        _activeFilters.Remove(AudioBookFilter.Completed);
+
+        await FilterAudiobookList();
+    }
+
+    private async void SelectAllFiltersCheckBox_OnChecked(object sender, RoutedEventArgs e)
+    {
+        InProgressFilterCheckBox.IsChecked =
+            NotStartedFilterCheckBox.IsChecked = CompletedFilterCheckBox.IsChecked = true;
+    }
+
+    private async void SelectAllFiltersCheckBox_OnUnchecked(object sender, RoutedEventArgs e)
+    {
+        InProgressFilterCheckBox.IsChecked =
+            NotStartedFilterCheckBox.IsChecked = CompletedFilterCheckBox.IsChecked = false;
+    }
+
+    private void SelectAllFiltersCheckBox_OnIndeterminate(object sender, RoutedEventArgs e)
+    {
+        if (InProgressFilterCheckBox.IsChecked == true && NotStartedFilterCheckBox.IsChecked == true &&
+            CompletedFilterCheckBox.IsChecked == true)
+            SelectAllFiltersCheckBox.IsChecked = false;
     }
 }